Offers

+++ service-center/src/main/java/com/maintainance/service_center/analytics/PerformanceSummaryResponse.java
@@ -5,19 +5,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class PerformanceSummaryResponse {
-    
+
     private long totalBookings;
     private long completedBookings;
     private long cancelledBookings;
     private double cancellationRate;
     private Double averageRating;
-    private BigDecimal totalRevenue;
+    private Double totalRevenue;
     private boolean revenueAvailable;
 }
